requirements.txt
fastapi
PyYaml
uvicorn
python-multipart